log4j-over-slf4j/src/test/java/org/apache/log4j/NDCTest.java
/**
 * Copyright (c) 2004-2011 QOS.ch
 * All rights reserved.
 *
 * Permission is hereby granted, free  of charge, to any person obtaining
 * a  copy  of this  software  and  associated  documentation files  (the
 * "Software"), to  deal in  the Software without  restriction, including
 * without limitation  the rights to  use, copy, modify,  merge, publish,
 * distribute,  sublicense, and/or sell  copies of  the Software,  and to
 * permit persons to whom the Software  is furnished to do so, subject to
 * the following conditions:
 *
 * The  above  copyright  notice  and  this permission  notice  shall  be
 * included in all copies or substantial portions of the Software.
 *
 * THE  SOFTWARE IS  PROVIDED  "AS  IS", WITHOUT  WARRANTY  OF ANY  KIND,
 * EXPRESS OR  IMPLIED, INCLUDING  BUT NOT LIMITED  TO THE  WARRANTIES OF
 * MERCHANTABILITY,    FITNESS    FOR    A   PARTICULAR    PURPOSE    AND
 * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
 * LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
 * OF CONTRACT, TORT OR OTHERWISE,  ARISING FROM, OUT OF OR IN CONNECTION
 * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 *
 */
package org.apache.log4j;

import junit.framework.TestCase;

/**
 * @author Ceki G&uuml;c&uuml;
 */
public class NDCTest extends TestCase {


  public void setUp() {
    assertEquals(0, NDC.getDepth());
  }

  public void tearDown() {
    NDC.clear();
  }

  public void testSmoke() {
    NDC.push("a");
    String back = NDC.pop();
    assertEquals("a", back);
  }

  public void testPop() {
    NDC.push("peek");
    String back = NDC.peek();
    assertEquals("peek", back);
  }

  public void testClear() {
    NDC.push("clear");
    NDC.clear();
    assertEquals(0, NDC.getDepth());
  }

}
