slf4j-ext/src/test/java/org/slf4j/NDCTest.java
/**
 * Copyright (c) 2004-2011 QOS.ch
 * All rights reserved.
 *
 * Permission is hereby granted, free  of charge, to any person obtaining
 * a  copy  of this  software  and  associated  documentation files  (the
 * "Software"), to  deal in  the Software without  restriction, including
 * without limitation  the rights to  use, copy, modify,  merge, publish,
 * distribute,  sublicense, and/or sell  copies of  the Software,  and to
 * permit persons to whom the Software  is furnished to do so, subject to
 * the following conditions:
 *
 * The  above  copyright  notice  and  this permission  notice  shall  be
 * included in all copies or substantial portions of the Software.
 *
 * THE  SOFTWARE IS  PROVIDED  "AS  IS", WITHOUT  WARRANTY  OF ANY  KIND,
 * EXPRESS OR  IMPLIED, INCLUDING  BUT NOT LIMITED  TO THE  WARRANTIES OF
 * MERCHANTABILITY,    FITNESS    FOR    A   PARTICULAR    PURPOSE    AND
 * NONINFRINGEMENT. IN NO EVENT SHALL THE AUTHORS OR COPYRIGHT HOLDERS BE
 * LIABLE FOR ANY CLAIM, DAMAGES OR OTHER LIABILITY, WHETHER IN AN ACTION
 * OF CONTRACT, TORT OR OTHERWISE,  ARISING FROM, OUT OF OR IN CONNECTION
 * WITH THE SOFTWARE OR THE USE OR OTHER DEALINGS IN THE SOFTWARE.
 *
 */
package org.slf4j;

import junit.framework.TestCase;

public class NDCTest extends TestCase {

  protected void setUp() throws Exception {
    super.setUp();
    MDC.clear();
  }

  protected void tearDown() throws Exception {
    super.tearDown();
  }
  
  public void testEmpty() {
    assertEquals("", NDC.pop());
  }
  
  public void testSmoke() {
    NDC.push("a");
    String result = NDC.pop();
    assertEquals("a",result);
  }

  public void testSmoke2() {
    NDC.push("a");
    NDC.push("b");
    String result1 = NDC.pop();
    String result0 = NDC.pop();
    assertEquals("b",result1);
    assertEquals("a",result0);
  }
}
